# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Сервис анализа патентной активности компаний.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/patents": {
      "get": {
        "tags": [
          "Patents"
        ],
        "summary": "List Patents",
        "description": "Получить список патентов с пагинацией и возможностью фильтрации.\n\nParameters:\n- **page** (int, optional): Номер страницы для пагинации. По умолчанию 1.\n- **pagesize** (int, optional): Количество патентов на странице. По умолчанию 10.\n- **filter_id** (int, optional): ID фильтра для фильтрации по списку ИНН владельцев.\n- **kind** (int, optional): Тип патента для фильтрации:\n    * 1 - Изобретение\n    * 2 - Полезная модель\n    * 3 - Промышленный образец\n- **actual** (bool, optional): Фильтр по актуальности патента:\n    * true - только действующие патенты\n    * false - только недействующие патенты\n\nReturns:\n- **PatentsList**: Объект, содержащий:\n    * total (int): Общее количество патентов\n    * items (List[Patent]): Список патентов с информацией о владельцах",
        "operationId": "list_patents_patents_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "pagesize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Pagesize"
            }
          },
          {
            "name": "filter_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "actual",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actual"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatentsList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Patents"
        ],
        "summary": "Create Patent",
        "description": "Создать новый патент.\n\nParameters:\n- **patent** (PatentCreate): Объект с данными для создания патента, содержащий:\n    * reg_number (int): Регистрационный номер патента\n    * reg_date (date): Дата регистрации\n    * kind (int): Тип патента (1 - Изобретение, 2 - Полезная модель, 3 - Промышленный образец)\n    * name (str): Название патента\n    * author_raw (str, optional): Список авторов\n    * owner_raw (str, optional): Список владельцев\n    * country_code (str, optional): Код страны правообладателя\n    * ... другие опциональные поля\n\nReturns:\n- **PatentDB**: Созданный патент",
        "operationId": "create_patent_patents_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatentDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patents/stats": {
      "get": {
        "tags": [
          "Patents"
        ],
        "summary": "Get Patents Stats",
        "description": "Получить статистику по патентам.\n\nParameters:\n- **filter_id** (int, optional): ID фильтра для получения статистики по конкретным ИНН.\n    ID фильтров начинаются с 4 и выше.\n\nReturns:\n- **PatentsStats**: Объект со статистикой, содержащий:\n    * total_patents (int): Общее количество патентов\n    * total_ru_patents (int): Количество российских патентов\n    * total_with_holders (int): Количество патентов с указанными владельцами\n    * total_ru_with_holders (int): Количество российских патентов с владельцами\n    * with_holders_percent (int): Процент патентов с указанными владельцами\n    * ru_with_holders_percent (int): Процент российских патентов с владельцами\n    * by_author_count (Dict): Распределение по количеству авторов\n    * by_patent_kind (Dict): Распределение по типам патентов",
        "operationId": "get_patents_stats_patents_stats_get",
        "parameters": [
          {
            "name": "filter_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatentsStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patents/{patent_kind}/{patent_reg_number}": {
      "get": {
        "tags": [
          "Patents"
        ],
        "summary": "Get Patent",
        "description": "Получить детальную информацию о патенте.\n\nParameters:\n- **patent_kind** (int): Тип патента:\n    * 1 - Изобретение\n    * 2 - Полезная модель\n    * 3 - Промышленный образец\n- **patent_reg_number** (int): Регистрационный номер патента\n\nReturns:\n- **PatentAdditionalFields**: Детальная информация о патенте, включая:\n    * Все стандартные поля патента\n    * owner_raw (str): Список владельцев одной строкой\n    * patent_holders (List[PatentHolder]): Структурированный список владельцев",
        "operationId": "get_patent_patents__patent_kind___patent_reg_number__get",
        "parameters": [
          {
            "name": "patent_kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Patent Kind"
            }
          },
          {
            "name": "patent_reg_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Patent Reg Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatentAdditionalFields"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Patents"
        ],
        "summary": "Update Patent",
        "description": "Обновить существующий патент.\n\nArgs:\n   patent_kind (int): вид патента.\n   patent_reg_number (int): регистрационный номер патента.\n   obj_in (PatentUpdate): данные для обновления патента.\n   session (AsyncSession): асинхронная сессия базы данных.\n\nReturns:\n   PatentDB: обновленный патент.",
        "operationId": "update_patent_patents__patent_kind___patent_reg_number__patch",
        "parameters": [
          {
            "name": "patent_kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Patent Kind"
            }
          },
          {
            "name": "patent_reg_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Patent Reg Number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatentDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Patents"
        ],
        "summary": "Delete Patent",
        "description": "Обновить существующий патент.\n\nParameters:\n- **patent_kind** (int): Тип патента:\n    * 1 - Изобретение\n    * 2 - Полезная модель\n    * 3 - Промышленный образец\n- **patent_reg_number** (int): Регистрационный номер патента\n- **obj_in** (PatentUpdate): Объект с данными для обновления:\n    * name (str, optional): Новое название патента\n    * actual (bool, optional): Статус актуальности\n    * author_raw (str, optional): Обновленный список авторов\n    * ... другие опциональные поля\n\nReturns:\n- **PatentDB**: Обновленный патент",
        "operationId": "delete_patent_patents__patent_kind___patent_reg_number__delete",
        "parameters": [
          {
            "name": "patent_kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Patent Kind"
            }
          },
          {
            "name": "patent_reg_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Patent Reg Number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/uploadfile/": {
      "post": {
        "tags": [
          "Patents"
        ],
        "summary": "Send Patent File",
        "description": "Загрузить Excel-файл с данными о патентах.\n\nParameters:\n- **file** (UploadFile): Excel-файл с данными о патентах.\n    Файл должен соответствовать установленному шаблону.\n\nReturns:\n- **Dict[str, str]**: Результат обработки файла:\n    * message (str): Сообщение о результате загрузки\n    * filename (str): Имя обработанного файла",
        "operationId": "send_patent_file_uploadfile__post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_send_patent_file_uploadfile__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patents/export": {
      "get": {
        "tags": [
          "Patents"
        ],
        "summary": "Export Patents",
        "description": "Экспортировать данные о патентах в формате XLSX.\n\nParameters:\n- **filter_id** (int, optional): ID фильтра для экспорта патентов по списку ИНН.\n    ID фильтров начинаются с 4 и выше.\n- **actual** (str, optional): Фильтр по актуальности:\n    * \"Актуально\" - только действующие патенты\n    * \"Неактуально\" - только недействующие патенты\n- **kind** (int, optional): Тип патента для фильтрации:\n    * 1 - Изобретение\n    * 2 - Полезная модель\n    * 3 - Промышленный образец\n\nReturns:\n- **StreamingResponse**: Поток с XLSX-файлом\n\nNote:\n- Без фильтров экспортируются первые 10000 патентов",
        "operationId": "export_patents_patents_export_get",
        "parameters": [
          {
            "name": "filter_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter Id"
            }
          },
          {
            "name": "actual",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actual"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_send_patent_file_uploadfile__post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_send_patent_file_uploadfile__post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatentCreate": {
        "properties": {
          "reg_number": {
            "type": "integer",
            "title": "Reg Number"
          },
          "reg_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reg Date"
          },
          "appl_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appl Date"
          },
          "owner_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Raw"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "actual": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "string"
              }
            ],
            "title": "Actual"
          },
          "subcategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategory"
          },
          "kind": {
            "type": "integer",
            "title": "Kind"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "appl_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appl Number"
          },
          "patent_starting_date": {
            "type": "string",
            "format": "date",
            "title": "Patent Starting Date"
          },
          "publication_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication Url"
          }
        },
        "type": "object",
        "required": [
          "reg_number",
          "name",
          "actual",
          "kind",
          "patent_starting_date"
        ],
        "title": "PatentCreate"
      },
      "PatentDB": {
        "properties": {
          "reg_number": {
            "type": "integer",
            "title": "Reg Number"
          },
          "reg_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reg Date"
          },
          "appl_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appl Date"
          },
          "owner_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Raw"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "actual": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "string"
              }
            ],
            "title": "Actual"
          },
          "subcategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategory"
          },
          "kind": {
            "type": "integer",
            "title": "Kind"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "appl_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appl Number"
          },
          "patent_starting_date": {
            "type": "string",
            "format": "date",
            "title": "Patent Starting Date"
          },
          "publication_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication Url"
          }
        },
        "type": "object",
        "required": [
          "reg_number",
          "name",
          "actual",
          "kind",
          "patent_starting_date"
        ],
        "title": "PatentDB"
      },
...
      "PatentsList": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PatentAdditionalFields"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "PatentsList"
      },
      "PatentsStats": {
        "properties": {
          "total_patents": {
            "type": "integer",
            "title": "Total Patents"
          },
          "total_ru_patents": {
            "type": "integer",
            "title": "Total Ru Patents"
          },
          "total_with_holders": {
            "type": "integer",
            "title": "Total With Holders"
          },
          "total_ru_with_holders": {
            "type": "integer",
            "title": "Total Ru With Holders"
          },
          "with_holders_percent": {
            "type": "integer",
            "title": "With Holders Percent"
          },
          "ru_with_holders_percent": {
            "type": "integer",
            "title": "Ru With Holders Percent"
          },
          "by_author_count": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Author Count"
          },
          "by_patent_kind": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Patent Kind"
          }
        },
        "type": "object",
        "required": [
          "total_patents",
          "total_ru_patents",
          "total_with_holders",
          "total_ru_with_holders",
          "with_holders_percent",
          "ru_with_holders_percent",
          "by_author_count",
          "by_patent_kind"
        ],
        "title": "PatentsStats"
      },
...
    }
  }
}